samantha-simple/vercel-deploy/requirements.txt
fastapi
uvicorn
google-genai==1.24.0
pydantic
